requirements_bookworm.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.12
    # via pyswitchbot
aiooui==0.1.9
    # via bluetooth-adapters
aiosignal==1.3.2
    # via aiohttp
attrs==25.1.0
    # via aiohttp
bleak==0.22.3
    # via
    #   bleak-retry-connector
    #   bluetooth-adapters
    #   pyswitchbot
bleak-retry-connector==3.9.0
    # via pyswitchbot
bluetooth-adapters==0.21.4
    # via bleak-retry-connector
certifi==2025.1.31
    # via
    #   influxdb-client
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
cryptography==44.0.1
    # via
    #   pyopenssl
    #   pyswitchbot
dbus-fast==2.33.0
    # via
    #   bleak
    #   bleak-retry-connector
    #   bluetooth-adapters
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via
    #   requests
    #   yarl
influxdb-client==1.48.0
    # via switchbot-dashboard (pyproject.toml)
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
pycparser==2.22
    # via cffi
pyopenssl==25.0.0
    # via pyswitchbot
pyswitchbot==0.56.0
    # via switchbot-dashboard (pyproject.toml)
python-dateutil==2.9.0.post0
    # via influxdb-client
python-dotenv==1.0.1
    # via switchbot-dashboard (pyproject.toml)
reactivex==4.0.4
    # via influxdb-client
requests==2.32.3
    # via switchbot-dashboard (pyproject.toml)
schedule==1.2.2
    # via switchbot-dashboard (pyproject.toml)
setuptools==75.8.0
    # via influxdb-client
six==1.17.0
    # via python-dateutil
typing-extensions==4.12.2
    # via reactivex
uart-devices==0.1.0
    # via bluetooth-adapters
urllib3==2.3.0
    # via
    #   influxdb-client
    #   requests
usb-devices==0.4.5
    # via bluetooth-adapters
yarl==1.18.3
    # via aiohttp
